Extracted overhead line tests into a new file
